moved slider
